Fixed GitHub dependencies
Optimized the process of installing dependencies from GitHub. It now
uses the tarball of the CURRENT commit in master. Previously it was
cloning the entire repository which wasn't needed and when it comes to
express and socket.io, things get pretty big. This also fixes some
issues that npm on Windows and other OSes have with malformed git
addresses.